move exception catcher to MRC
